@@ -1,16 +1,16 @@
 name: msal_js
 description: "A Dart wrapper for the 'Microsoft Authentication Library for JavaScript (MSAL.js)'."
-version: 1.3.0
+version: 1.4.0-nullsafety.0
 homepage: https://github.com/milestonetg/msal-js-dart
 repository: https://github.com/milestonetg/msal-js-dart
 issue_tracker: https://github.com/milestonetg/msal-js-dart/issues
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
-  js: ^0.6.2
+  js: 0.6.3-nullsafety.3
 
 dev_dependencies:
-  pedantic: ^1.9.2
-  test: ^1.15.0
+  pedantic: 1.10.0-nullsafety.3
+  test: 1.16.0-nullsafety.17
