@@ -2,39 +2,46 @@ part of '../msal_js.dart';
 
 /// Used to configure a [UserAgentApplication].
 class Configuration {
-  AuthOptions get auth =>
-      _jsObject.auth == null ? null : AuthOptions._fromJsObject(_jsObject.auth);
+  AuthOptions? get auth {
+    final auth = _jsObject.auth;
+    return auth == null ? null : AuthOptions._fromJsObject(auth);
+  }
 
   /// Sets the auth options.
-  set auth(AuthOptions value) => _jsObject.auth = value._jsObject;
+  set auth(AuthOptions? value) => _jsObject.auth = value?._jsObject;
 
-  CacheOptions get cache => _jsObject.cache == null
-      ? null
-      : CacheOptions._fromJsObject(_jsObject.cache);
+  CacheOptions? get cache {
+    final cache = _jsObject.cache;
+    return cache == null ? null : CacheOptions._fromJsObject(cache);
+  }
 
   /// Sets the cache options.
-  set cache(CacheOptions value) => _jsObject.cache = value._jsObject;
+  set cache(CacheOptions? value) => _jsObject.cache = value?._jsObject;
 
-  SystemOptions get system => _jsObject.system == null
-      ? null
-      : SystemOptions._fromJsObject(_jsObject.system);
+  SystemOptions? get system {
+    final system = _jsObject.system;
+    return system == null ? null : SystemOptions._fromJsObject(system);
+  }
 
   /// Sets the library specific options.
-  set system(SystemOptions value) => _jsObject.system = value._jsObject;
+  set system(SystemOptions? value) => _jsObject.system = value?._jsObject;
 
   @Deprecated('''
     FrameworkOptions is only used by the Angular MSAL libraries and will do nothing in this package. 
     See [FrameworkOptions] for more information.
   ''')
-  FrameworkOptions get framework => _jsObject.framework == null
-      ? null
-      : FrameworkOptions._fromJsObject(_jsObject.framework);
+  FrameworkOptions? get framework {
+    final framework = _jsObject.framework;
+    return framework == null ? null : FrameworkOptions._fromJsObject(framework);
+  }
+
   @Deprecated('''
     FrameworkOptions is only used by the Angular MSAL libraries and will do nothing in this package. 
     See [FrameworkOptions] for more information.
   ''')
-  set framework(FrameworkOptions value) =>
-      _jsObject.framework = value._jsObject;
+  set framework(FrameworkOptions? value) {
+    _jsObject.framework = value?._jsObject;
+  }
 
   final interop.Configuration _jsObject;
 
@@ -45,12 +52,12 @@ class Configuration {
 
 /// Used to configure the auth options for a [UserAgentApplication].
 class AuthOptions {
-  String get clientId => _jsObject.clientId;
+  String? get clientId => _jsObject.clientId;
 
   /// Sets the client ID of the registered application.
-  set clientId(String value) => _jsObject.clientId = value;
+  set clientId(String? value) => _jsObject.clientId = value;
 
-  String get authority => _jsObject.authority;
+  String? get authority => _jsObject.authority;
 
   /// Sets the authority to be used.
   ///
@@ -61,26 +68,28 @@ class AuthOptions {
   ///   the TenantID property of the directory)
   /// - In Azure B2C, it is of the form https://<instance>/tfp/<tenant>/<policyName>/
   /// - Defaults to https://login.microsoftonline.com/common.
-  set authority(String value) => _jsObject.authority = value;
+  set authority(String? value) => _jsObject.authority = value;
 
-  bool get validateAuthority => _jsObject.validateAuthority;
+  bool? get validateAuthority => _jsObject.validateAuthority;
 
   /// Sets whether the authority should be validated.
   ///
   /// When set to `true` (default), MSAL will compare the application's authority against well-known URL
   /// templates representing well-formed authorities. It is useful when the authority is obtained at
   /// run time to prevent MSAL from displaying authentication prompts from malicious pages.
-  set validateAuthority(bool value) => _jsObject.validateAuthority = value;
+  set validateAuthority(bool? value) => _jsObject.validateAuthority = value;
 
-  List<String> get knownAuthorities =>
-      jsDecodeList<String>(_jsObject.knownAuthorities);
+  List<String>? get knownAuthorities {
+    return jsDecodeList<String>(_jsObject.knownAuthorities);
+  }
 
   /// If [validateAuthority] is set to true, this will be used to set the
   /// Trusted Host list.
   ///
   /// Defaults to empty.
-  set knownAuthorities(List<String> value) =>
-      _jsObject.knownAuthorities = jsEncode(value);
+  set knownAuthorities(List<String>? value) {
+    return _jsObject.knownAuthorities = jsEncode(value);
+  }
 
   dynamic get redirectUri => _jsObject.redirectUri;
 
@@ -124,16 +133,17 @@ class AuthOptions {
     _jsObject.postLogoutRedirectUri = value;
   }
 
-  bool get navigateToLoginRequestUrl => _jsObject.navigateToLoginRequestUrl;
+  bool? get navigateToLoginRequestUrl => _jsObject.navigateToLoginRequestUrl;
 
   /// Sets whether the user will be redirected to the URL that initiated the login
   /// after a successful login.
   ///
   /// Defaults to `true`.
-  set navigateToLoginRequestUrl(bool value) =>
-      _jsObject.navigateToLoginRequestUrl = value;
+  set navigateToLoginRequestUrl(bool? value) {
+    _jsObject.navigateToLoginRequestUrl = value;
+  }
 
-  String get authorityMetadata => _jsObject.authorityMetadata;
+  String? get authorityMetadata => _jsObject.authorityMetadata;
 
   /// Sets the OpenID configuration metadata for the configured authority.
   ///
@@ -141,7 +151,7 @@ class AuthOptions {
   ///
   /// See https://github.com/AzureAD/microsoft-authentication-library-for-js/blob/dev/lib/msal-core/docs/performance.md#bypass-authority-metadata-resolution
   /// for more information.
-  set authorityMetadata(String value) => _jsObject.authorityMetadata = value;
+  set authorityMetadata(String? value) => _jsObject.authorityMetadata = value;
 
   final interop.AuthOptions _jsObject;
 
@@ -152,8 +162,10 @@ class AuthOptions {
 
 /// Used to configure the cache options for a [UserAgentApplication].
 class CacheOptions {
-  CacheLocation get cacheLocation =>
-      _stringToCacheLocation(_jsObject.cacheLocation);
+  CacheLocation? get cacheLocation {
+    final cacheLocation = _jsObject.cacheLocation;
+    return cacheLocation == null ? null : _stringToCacheLocation(cacheLocation);
+  }
 
   /// Sets which browser cache should be used.
   ///
@@ -161,10 +173,12 @@ class CacheOptions {
   /// across multiple browser tabs.
   ///
   /// Defaults to [CacheLocation.sessionStorage].
-  set cacheLocation(CacheLocation value) =>
-      _jsObject.cacheLocation = _cacheLocationToString(value);
+  set cacheLocation(CacheLocation? value) {
+    _jsObject.cacheLocation =
+        value == null ? null : _cacheLocationToString(value);
+  }
 
-  bool get storeAuthStateInCookie => _jsObject.storeAuthStateInCookie;
+  bool? get storeAuthStateInCookie => _jsObject.storeAuthStateInCookie;
 
   /// Sets whether the authentication state should be stored in a cookie.
   ///
@@ -172,8 +186,9 @@ class CacheOptions {
   /// flows in the browser cookies.
   ///
   /// Defaults to `false`.
-  set storeAuthStateInCookie(bool value) =>
-      _jsObject.storeAuthStateInCookie = value;
+  set storeAuthStateInCookie(bool? value) {
+    _jsObject.storeAuthStateInCookie = value;
+  }
 
   final interop.CacheOptions _jsObject;
 
@@ -184,35 +199,38 @@ class CacheOptions {
 
 /// Used to configure library specific options for a [UserAgentApplication].
 class SystemOptions {
-  Logger get logger =>
-      _jsObject.logger == null ? null : Logger._fromJsObject(_jsObject.logger);
+  Logger? get logger {
+    final logger = _jsObject.logger;
+    return logger == null ? null : Logger._fromJsObject(logger);
+  }
 
   /// Sets the logger to be used by the application.
   ///
   /// Defaults to `null`.
-  set logger(Logger value) => _jsObject.logger = value._jsObject;
+  set logger(Logger? value) => _jsObject.logger = value?._jsObject;
 
-  num get loadFrameTimeout => _jsObject.loadFrameTimeout;
+  num? get loadFrameTimeout => _jsObject.loadFrameTimeout;
 
   /// Sets the maximum time in milliseconds the library will wait for an iframe to load.
   ///
   /// Defaults to 6000.
-  set loadFrameTimeout(num value) => _jsObject.loadFrameTimeout = value;
+  set loadFrameTimeout(num? value) => _jsObject.loadFrameTimeout = value;
 
-  num get tokenRenewalOffsetSeconds => _jsObject.tokenRenewalOffsetSeconds;
+  num? get tokenRenewalOffsetSeconds => _jsObject.tokenRenewalOffsetSeconds;
 
   /// Sets the window of offset (in seconds) needed to renew the token before expiry.
   ///
   /// Defaults to 300.
-  set tokenRenewalOffsetSeconds(num value) =>
-      _jsObject.tokenRenewalOffsetSeconds = value;
+  set tokenRenewalOffsetSeconds(num? value) {
+    _jsObject.tokenRenewalOffsetSeconds = value;
+  }
 
-  num get navigateFrameWait => _jsObject.navigateFrameWait;
+  num? get navigateFrameWait => _jsObject.navigateFrameWait;
 
   /// Sets the wait time (in milliseconds) for hidden iframe navigation.
   ///
   /// Defaults to 500.
-  set navigateFrameWait(num value) => _jsObject.navigateFrameWait = value;
+  set navigateFrameWait(num? value) => _jsObject.navigateFrameWait = value;
 
   final interop.SystemOptions _jsObject;
 
@@ -228,15 +246,21 @@ class SystemOptions {
   This should not be used and will be removed in the future.
 ''')
 class FrameworkOptions {
-  List<String> get unprotectedResources =>
-      jsDecodeList<String>(_jsObject.unprotectedResources);
-  set unprotectedResources(List<String> value) =>
-      _jsObject.unprotectedResources = jsEncode(value);
+  List<String>? get unprotectedResources {
+    return jsDecodeList<String>(_jsObject.unprotectedResources);
+  }
 
-  Map<String, List<String>> get protectedResourceMap =>
-      jsDecodeMap<List<String>>(_jsObject.protectedResourceMap);
-  set protectedResourceMap(Map<String, List<String>> value) =>
-      _jsObject.protectedResourceMap = jsEncode(value);
+  set unprotectedResources(List<String>? value) {
+    _jsObject.unprotectedResources = jsEncode(value);
+  }
+
+  Map<String, List<String>>? get protectedResourceMap {
+    return jsDecodeMap<List<String>>(_jsObject.protectedResourceMap);
+  }
+
+  set protectedResourceMap(Map<String, List<String>>? value) {
+    _jsObject.protectedResourceMap = jsEncode(value);
+  }
 
   final interop.FrameworkOptions _jsObject;
 
