@@ -9,8 +9,8 @@ typedef RedirectUriCallback = String Function();
 ///
 /// - [error] - If not null, contains the error that caused authentication to fail.
 /// - [response] - If not null, contains the successful authentication response data.
-typedef AuthResponseCallback = void Function(AuthException error,
-    [AuthResponse response]);
+typedef AuthResponseCallback = void Function(AuthException? error,
+    [AuthResponse? response]);
 
 /// An MSAL authentication context.
 ///
@@ -18,7 +18,7 @@ typedef AuthResponseCallback = void Function(AuthException error,
 /// for more information.
 class UserAgentApplication {
   /// Gets the authority for this application.
-  String get authority => _jsObject.authority;
+  String? get authority => _jsObject.authority;
 
   /// Sets the authority of this application.
   ///
@@ -29,7 +29,7 @@ class UserAgentApplication {
   ///   the TenantID property of the directory)
   /// - In Azure B2C, it is of the form https://<instance>/tfp/<tenant>/<policyName>/
   /// - Default value is: "https://login.microsoftonline.com/common"
-  set authority(String value) => _jsObject.authority = value;
+  set authority(String? value) => _jsObject.authority = value;
 
   final interop.UserAgentApplication _jsObject;
 
@@ -37,8 +37,6 @@ class UserAgentApplication {
   ///
   /// If the [configuration] is invalid, a [ClientConfigurationException] will be thrown.
   factory UserAgentApplication(Configuration configuration) {
-    if (configuration == null) throw ArgumentError.notNull('configuration');
-
     try {
       return UserAgentApplication._fromJsObject(
           interop.UserAgentApplication(configuration._jsObject));
@@ -53,14 +51,15 @@ class UserAgentApplication {
   ///
   /// This callback must be set before [acquireTokenRedirect] and [loginRedirect] can be used.
   void handleRedirectCallback(AuthResponseCallback callback) {
-    if (callback == null) throw ArgumentError.notNull('callback');
-
-    void jsCallback(interop.AuthError error, [interop.AuthResponse response]) {
-      callback(convertJsAuthError(error), AuthResponse._fromJsObject(response));
+    void jsCallback(interop.AuthError? error,
+        [interop.AuthResponse? response]) {
+      callback(error == null ? null : convertJsAuthError(error),
+          response == null ? null : AuthResponse._fromJsObject(response));
     }
 
-    _callJsMethod(
-        () => _jsObject.handleRedirectCallback(allowInterop(jsCallback)));
+    _callJsMethod(() {
+      _jsObject.handleRedirectCallback(allowInterop(jsCallback));
+    });
   }
 
   /// Acquires an access token using interactive authentication via a popup Window.
@@ -104,24 +103,23 @@ class UserAgentApplication {
   Future<AuthResponse> ssoSilent(AuthRequest request) async {
     final response = await _convertMsalPromise<interop.AuthResponse>(
         _callJsMethod(() => _jsObject.ssoSilent(request._jsObject)));
-
     return AuthResponse._fromJsObject(response);
   }
 
   /// Gets all currently cached unique accounts based on `homeAccountIdentifier`.
   List<Account> getAllAccounts() {
-    return _callJsMethod(() => _jsObject
-        .getAllAccounts()
-        .cast<interop.Account>()
-        .map((jsAccount) => Account._fromJsObject(jsAccount))
-        .toList());
+    return _callJsMethod(() {
+      return (_jsObject.getAllAccounts() ?? [])
+          .cast<interop.Account>()
+          .map((jsAccount) => Account._fromJsObject(jsAccount))
+          .toList();
+    });
   }
 
   /// Gets the signed in account or `null` if no-one is signed in.
-  Account getAccount() {
+  Account? getAccount() {
     return _callJsMethod(() {
       final jsAccount = _jsObject.getAccount();
-
       return jsAccount == null ? null : Account._fromJsObject(jsAccount);
     });
   }
@@ -129,7 +127,7 @@ class UserAgentApplication {
   /// Initiates the login process by opening a popup browser window.
   ///
   /// Will throw an [AuthException] on failure.
-  Future<AuthResponse> loginPopup([AuthRequest request]) async {
+  Future<AuthResponse> loginPopup([AuthRequest? request]) async {
     final response = await _convertMsalPromise<interop.AuthResponse>(
         _callJsMethod(() => _jsObject.loginPopup(request?._jsObject)));
 
@@ -137,44 +135,44 @@ class UserAgentApplication {
   }
 
   /// Initiates the login process by redirecting the user to the authorization endpoint.
-  void loginRedirect([AuthRequest request]) {
-    return _callJsMethod(() => _jsObject.loginRedirect(request?._jsObject));
+  void loginRedirect([AuthRequest? request]) {
+    _callJsMethod(() => _jsObject.loginRedirect(request?._jsObject));
   }
 
   /// Logs out the current user, and redirects to the `postLogoutRedirectUri`.
-  void logout() {
-    _callJsMethod(() => _jsObject.logout());
-  }
+  void logout() => _callJsMethod(_jsObject.logout);
 
   /// Sets the logger to be used by the application.
-  void setLogger(Logger logger) {
+  void setLogger(Logger? logger) {
     _callJsMethod(() => _jsObject.setLogger(logger?._jsObject));
   }
 
   /// Returns whether a login is currently in progress.
-  bool getLoginInProgress() {
+  bool? getLoginInProgress() {
     return _callJsMethod(() => _jsObject.getLoginInProgress());
   }
 
   /// Returns the current configuration of this user agent application.
-  Configuration getCurrentConfiguration() {
-    return _callJsMethod(
-        () => Configuration._fromJsObject(_jsObject.getCurrentConfiguration()));
+  Configuration? getCurrentConfiguration() {
+    return _callJsMethod(() {
+        final configuration = _jsObject.getCurrentConfiguration();
+        return configuration == null
+            ? null
+            : Configuration._fromJsObject(configuration);
+      });
   }
 
   /// Returns the post-logout redirect URI currently configured.
   ///
   /// If the post-logout redirect URI was configured as a function, it
   /// will be evaluated and its result will be returned.
-  String getPostLogoutRedirectUri() {
-    return _callJsMethod(() => _jsObject.getPostLogoutRedirectUri());
+  String? getPostLogoutRedirectUri() {
+    return _callJsMethod(_jsObject.getPostLogoutRedirectUri);
   }
 
   /// Returns the redirect URI currently configured.
   ///
   /// If the redirect URI was configured as a function, it
   /// will be evaluated and its result will be returned.
-  String getRedirectUri() {
-    return _callJsMethod(() => _jsObject.getRedirectUri());
-  }
+  String? getRedirectUri() => _callJsMethod(_jsObject.getRedirectUri);
 }
