@@ -25,27 +25,27 @@ void main() {
 
     // Assert config can be read back
     void assertConfig() {
-      expect(config.auth.authority,
+      expect(config.auth!.authority,
           equals('https://login.microsoftonline.com/common'));
-      expect(config.auth.clientId, equals('fakeid'));
-      expect(config.auth.navigateToLoginRequestUrl, isFalse);
-      expect(config.auth.postLogoutRedirectUri, equals('post'));
-      expect(config.auth.redirectUri, equals('redirect'));
-      expect(config.auth.validateAuthority, isFalse);
+      expect(config.auth!.clientId, equals('fakeid'));
+      expect(config.auth!.navigateToLoginRequestUrl, isFalse);
+      expect(config.auth!.postLogoutRedirectUri, equals('post'));
+      expect(config.auth!.redirectUri, equals('redirect'));
+      expect(config.auth!.validateAuthority, isFalse);
       expect(
-          config.auth.knownAuthorities, equals(['login.microsoftonline.com']));
-      expect(config.cache.cacheLocation, equals(CacheLocation.localStorage));
-      expect(config.cache.storeAuthStateInCookie, isTrue);
-      expect(config.system.loadFrameTimeout, equals(1));
-      expect(config.system.navigateFrameWait, equals(1));
-      expect(config.system.tokenRenewalOffsetSeconds, equals(1));
+          config.auth!.knownAuthorities, equals(['login.microsoftonline.com']));
+      expect(config.cache!.cacheLocation, equals(CacheLocation.localStorage));
+      expect(config.cache!.storeAuthStateInCookie, isTrue);
+      expect(config.system!.loadFrameTimeout, equals(1));
+      expect(config.system!.navigateFrameWait, equals(1));
+      expect(config.system!.tokenRenewalOffsetSeconds, equals(1));
     }
 
     assertConfig();
 
     // Create user agent application and get config from there
     final userAgentApp = UserAgentApplication(config);
-    config = userAgentApp.getCurrentConfiguration();
+    config = userAgentApp.getCurrentConfiguration()!;
 
     // Assert config is the same after retrieved from user agent app
     assertConfig();
